Add simulation_mutex class
This class protects the entire simulation with a mutex.
It provides a method for locking the mutex.

It also provides getters and setters that bypass the mutex for several of its members, as they are accessed with std::atomic_ref instead.

This replaces the previous strategy of locking the world, instead of the whole simulation.