WIP: Remove duplicate getters and setters
The getters and setters for some of the members of simulation existed in both simulation and simulation_mutex.
This replaces both sets with a single set of accessors that return std::atomic_ref<T>.
This also allows atomically performing RMW operations on the members.
Before this, that was done with set_field(get_field() + 1) or similar, which was non-atomic.